requirements.txt
numpy
torch
transformers
tqdm
pillow
scikit-learn
pandas
sentence-transformers
